+++ test/bash/LNX/install-terraform-docs_01.sh
@@ -1,0 +1,9 @@
+#!/usr/bin/env bash
+
+if [[ "$DEV_TEST_BASH_CONTAINER_ENVIRONMENT" != 'ON' ]]; then
+  echo 'Error: invalid test environment'
+  exit 1
+fi
+
+export INST64_SYSTEM_WIDE='YES'
+sudo -E /source/install-terraform-docs