+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Trivy
+  - Multi Machine support
 - IstioCTL
   - Multi Machine support
 - SHFMT